remove src/utils/background files; move getAspectRatios to src/utils/background/helpers
