--- src/utils/background/globals.js
@@ -1,6 +1,0 @@
-// TODO: probably not needed anymore
-export let runtimePort;
-
-export let connection; // RTCMultiConnection
-
-export let videoPlayers = [];
